fix strange partials
